259 feat 필터 기반 식당 조회 메서드 추가

+++ src/main/java/com/sparta/goodbite/domain/review/dto/ReviewResponseDto.java
@@ -5,7 +5,7 @@ import com.sparta.goodbite.domain.review.entity.WaitingReview;
 import java.time.LocalDateTime;
 
 public record ReviewResponseDto(Long reviewId, String nickname, Long restaurantId,
-                                String restaurantName, float rating, String content,
+                                String restaurantName, double rating, String content,
                                 LocalDateTime createdAt) {
 
     public static ReviewResponseDto from(ReservationReview review) {
